requirements.txt
Django==4.0.3
django-audiofield==0.10.0
django-cors-headers==3.7.0
django-debug-toolbar==2.0
django-environ==0.8.1
django-filter==21.1
django-imagekit==4.1.0
django-jazzmin==2.4.9
django-modeltranslation==0.17.3
django-mptt==0.13.4
django-summernote==0.8.20.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
Pillow==8.2.0
psycopg2-binary==2.9.3
python-slugify==6.1.1
django-solo==2.0.0
djangorestframework-recursive
django-audiofield==0.10.0
django-environ==0.8.1
python-dateutil==2.8.2
django-tinymce==3.4.0
XlsxWriter==3.0.3 
beautifulsoup4==4.11.1
urllib3==1.26.9